Fix: rename Signup, UserController
