+++ project/src/main/java/com/favorite/project/User/RestController/UserController.java
@@ -1,32 +1,40 @@
 package com.favorite.project.User.RestController;
 
-//import com.favorite.project.mapper.UserMapper;
-
-import com.favorite.project.User.UserDataService;
 import com.favorite.project.User.UserService;
-import com.favorite.project.User.dto.UserDTO;
+import com.favorite.project.User.domain.User;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Slf4j
-@RestController
+import java.util.List;
+
 @RequiredArgsConstructor
+@RestController
 @RequestMapping(value = "/users")
 public class UserController {
-
-    private Logger logger = LoggerFactory.getLogger(getClass());
-
     private final UserService userService;
 
 
-    @PostMapping("/signup")
-    public boolean signup(@RequestBody UserDTO userDTO) {
-        return userService.insertOneUser(userDTO);
+    //json객체
+
+    @GetMapping
+    public List<User> getAllUsers() {
+
+        return userService.select();
+
 
     }
+
+
+//    //뷰로 가져오기
+//
+//    @GetMapping("/showUser")
+//    public ModelAndView showUser() {
+//        ModelAndView modelAndView = new ModelAndView();
+//        modelAndView.setViewName("userList");
+//        return modelAndView;
+//    }
 
 
 }
